Move and rename existing training and quals e2e tests
